Restructure profiler, allow named build directory (#139)
* Add option for builddir to spiralprofiler; restructure command files

* Update README; add ability to turn on keeptemp within Spiral

+++ profiler/targets/darwin-x86/matrix.cmd
@@ -1,19 +1,22 @@
 #!/bin/sh
 
-#  Copyright (c) 2018-2021, Carnegie Mellon University
+#  Copyright (c) 2018-2025, Carnegie Mellon University
 #  See LICENSE for details
 
 ##  Use cmake to build the project (PROJECT=matrix) for C language (SUFFIX=c) 
 
-TEMPDIR=$PWD
-cp -f ../../targets/common/CMakeLists.txt $TEMPDIR/CMakeLists.txt
-rm -rf build && mkdir build && cd build
-cmake -DPROJECT:STRING=matrix -DSUFFIX:STRING=c -DEXFLAGS:STRING=-march\=native .. > /dev/null
-make install > /dev/null
-cd ..
-
-if [ -f ./matrix ]; then
-    ./matrix > matrix.txt
+if [ "$1" = "build" ]; then
+    ##  Build the code
+    rm -rf build && mkdir build && cd build
+    cmake -DPROJECT:STRING=matrix -DSUFFIX:STRING=c -DEXFLAGS:STRING=-march\=native .. > /dev/null
+    make install > /dev/null
+    cd ..
 else
-    touch matrix.txt
+    ##  Run the code
+    if [ -f ./matrix ]; then
+        ./matrix > matrix.txt
+    else
+        touch matrix.txt
+    fi
 fi
+
